+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 1.0.3
 - Added an option to specify the custom colour for non-player/system messages' title (part before the comma) in chat.
 - Injected a silly game difficulty indicator to the lobby browser, likely phased out by the DAHM 1.16's native lobby ordering.